Support both sql.DB and sql.Tx
Replace *sql.DB with custom dbOrTx interface.
If the user passes sql.DB, start and commit a transaction automatically.
If the user passes sql.Tx, do not rollback or commit the transaction.

This allows the user to add custom logic into a single transaction. For
example, update an oid table T1 after Import. If T1 is used to GC, it
will not race with Import like previous code.